Refine REAPER_JAR search
Do not search from '/' to avoid log cluttering and
save startup time.

And anyway, searching from '/' kills the very idea
of searching for a *user* JAR (as apposed to a
system JAR) and also makes the search result less
predictable (as *all* JARs will eventually be
found).

The main motivation was that systemd services start
with '/' as the working directory, so the system log
gets filled with errors on each service start.